Allow segment to be specified as player prop
This avoids the need to set the segment directly on the underlying
player, using the ref.

We had a problem in the teachers app where segments weren't shareable.
It turned out that the code in this repo was responding to video URI
changes by setting the URI in the player, but clearing the segment in
the process.

[#174493760]

Signed-off-by: Alex Glen <alex@boclips.com>